requirements.txt
slack-sansio[aiohttp]
PyYAML
aiohttp